% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los.R
\name{los}
\alias{los}
\title{List Objects with Sizes (Shortened Name: los)}
\usage{
los(envir = .GlobalEnv, order = "decreasing")
}
\arguments{
\item{envir}{Environment to search for objects, defaults to global environment.}

\item{order}{Descending or ascending order for size sorting. Default is 'decreasing'.}
}
\value{
Prints a table of object names and their formatted sizes. Does not return any value.
}
\description{
This function \code{los} lists all objects in the specified environment, sorted by their memory sizes in a human-readable format. It allows for quick identification of large objects that may be using significant memory.
}
\examples{
los()
los(envir = .GlobalEnv, order = "decreasing")
}
